requirements.txt
fastapi>=0.111.0,<0.112.0
nost_tools >=2.3.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.1
PyJWT[crypto]>=2.8.0
httpx>=0.27.0
cachetools>=5.3.0